requirements.txt
# Kirigami 画像→PPTX 変換ツール (Python 3.10–3.12)
# 仮想環境: python -m venv .venv && source .venv/bin/activate

# 画像・PDF
Pillow>=10.0.0
opencv-python-headless>=4.8.0
PyMuPDF>=1.23.0

# レイヤー分解
fal-client>=0.4.0

# OCR
paddlepaddle>=2.5.0
paddleocr>=2.7.0
google-generativeai>=0.3.0

# LLM 補正
anthropic>=0.18.0
google-generativeai>=0.3.0
requests>=2.28.0  # Ollama ローカル呼び出し

# 出力
python-pptx>=0.6.21

# 設定
PyYAML>=6.0
python-dotenv>=1.0.0

# CLI / Web UI
Gradio>=4.0.0

# オプション: CPU レイヤー分解（ローカル推論）
# 未使用時はインストール不要
# torch>=2.0.0
# diffusers>=0.25.0
# transformers>=4.35.0
